requirements.txt
inquirer
nicegui
pyjwt